+++ core/search/wildcard_filter.cpp
@@ -27,28 +27,10 @@
 #include "multiterm_query.hpp"
 #include "term_query.hpp"
 #include "index/index_reader.hpp"
-#include "utils/wildcard_utils.hpp"
 #include "utils/automaton_utils.hpp"
 #include "utils/hash_utils.hpp"
 
 NS_ROOT
-
-inline irs::bytes_ref unescape(const irs::bytes_ref& in, irs::bstring& out) {
-  out.reserve(in.size());
-
-  bool copy = true;
-  std::copy_if(in.begin(), in.end(), std::back_inserter(out),
-               [&copy](irs::byte_type c) {
-    if (c == irs::WildcardMatch::ESCAPE) {
-      copy = !copy;
-    } else {
-      copy = true;
-    }
-    return copy;
-  });
-
-  return out;
-}
 
 DEFINE_FILTER_TYPE(by_wildcard)
 DEFINE_FACTORY_DEFAULT(by_wildcard)
